Add case management API and metrics
