+++ apps/legal_discovery/src/components/MetricCard.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+function MetricCard({ icon, label, value }) {
+  return (
+    <div className="metric-card">
+      <i className={`fa ${icon} text-xl mb-1`}></i>
+      <span className="value">{value}</span>
+      <span className="label">{label}</span>
+    </div>
+  );
+}
+export default MetricCard;
