+++ apps/legal_discovery/src/utils.js
@@ -1,0 +1,7 @@
+export function fetchJSON(url, options) {
+  return fetch(url, options).then(r => r.json());
+}
+
+export function alertResponse(d) {
+  alert(d.message || 'Done');
+}
