+++ apps/legal_discovery/static/style.css
@@ -133,6 +133,9 @@ h2 {
     color: var(--text-colour);
     cursor: pointer;
     font-weight: 500;
+    display: inline-flex;
+    align-items: center;
+    gap: 4px;
 }
 
 .tab-button.active {
@@ -445,21 +448,48 @@ progress::-webkit-progress-value {
 
 /* Team pipeline */
 .pipeline {
-    display: flex;
-    flex-wrap: wrap;
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(120px, 1fr));
     gap: 12px;
 }
 .pipeline .stage {
-    flex: 1 1 120px;
     background: var(--chat-bg);
     border: 1px solid var(--border-colour);
     border-radius: 8px;
     padding: 10px;
     text-align: center;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 4px;
 }
-.pipeline .stage span:last-child {
-    display: block;
+.pipeline .stage i {
+    color: var(--primary-color);
+    font-size: 1.4em;
+}
+.pipeline .stage .count {
     font-size: 1.1em;
     font-weight: bold;
-    margin-top: 4px;
+}
+
+/* Dashboard metrics */
+.metrics-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(100px, 1fr));
+    gap: 12px;
+    margin-bottom: 8px;
+}
+.metric-card {
+    background: var(--chat-bg);
+    border: 1px solid var(--border-colour);
+    border-radius: 8px;
+    padding: 8px;
+    text-align: center;
+}
+.metric-card .value {
+    font-size: 1.4em;
+    font-weight: bold;
+}
+.metric-card .label {
+    color: var(--subtitle-colour);
 }
