+++ apps/legal_discovery/templates/dashboard.html
@@ -18,10 +18,7 @@
     <script src="{{ url_for('static', filename='vis-timeline/vis-timeline-graph2d.min.js') }}"></script>
     <!-- React and Socket.IO -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.0.1/socket.io.js"></script>
-    <script crossorigin src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
-    <script src="https://unpkg.com/babel-standalone@7/babel.min.js"></script>
-    <script type="text/babel" src="{{ url_for('static', filename='dashboard-react.jsx') }}" defer></script>
+    <script type="module" src="{{ url_for('static', filename='bundle.js') }}" defer></script>
 </head>
 <body>
     <nav>
